+++ runner/rules/CHANGELOG.md
@@ -10,6 +10,9 @@
 
 **Breaking Changes**
 
+* `ServiceTestRule.startService` will now throw if the provided intent does not
+  launch a service.
+
 **API Changes**
 
 **Breaking API Changes**